HLA-10 dodajanje-datoteke-learn.py-ter-spreminjanje-kode-v-recognize.py
